Fix: user can comment on their own reservation
